+++ test/IntegrationTests/UnitTest1.cs
@@ -1,0 +1,14 @@
+using System;
+using Xunit;
+
+namespace IntegrationTests
+{
+    public class UnitTest1
+    {
+        [Fact]
+        public void Test1()
+        {
+
+        }
+    }
+}
